data-source-app/frontend/requirements.txt
Flask==2.3.3
psycopg[binary]==3.1.12
PyYAML==6.0.1
